requirements.txt
qiskit>=1.0.0
qiskit-aer>=0.13.0
qiskit-ibm-runtime>=0.20.0
torch>=2.0.0
torch-geometric>=2.4.0
networkx>=3.0
anthropic>=0.18.0
stable-baselines3>=2.1.0
numpy>=1.24.0
matplotlib>=3.7.0
python-dotenv>=1.0.0
pytest>=7.4.0
